requirements.txt
python-dotenv
requests
beautifulsoup4
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
streamlit
numpy==1.26.4
torch==2.2.0
sentence-transformers
